MAJOR UPDATE. Numerous deep revisions based on new approach to validation. See below.
Because of the two types of inclusion in TAN files (@include and @which), validation needs
to be made not only on the file but upon the fragments adopted from other files. This means
that it is easier to expand the original document through a chain of XSLT stylesheets that mark
where errors occur. Schematron validation, then, when evaluating an element, merely finds its
path and looks for the corresponding path in the document transformed for errors, looking for
<error>, <help>, <warning>, and <message> children, and acting on them when found. The global
variable that is key in this case is $self-prepped, along with $sources-prepped, in the case of
class-2 files.

One extra advantage to taking this approach is that the resultant $self-prepped is quite close
to an ideal model for post-processing stylesheet needs, thereby saving a lot of work for those
who wish to work with TAN files.

This commit reflects the first phase in the conversion out of Schematron. A great amount of old
code is retained, sometimes commented out, be pruned later, after refinement, testing, and
documentation.